#645: display games like in browse
