+++ packages/gdl-frontend/routes.js
@@ -61,9 +61,6 @@ routes.add('game', `/${langParam}/games/details/:id`, 'games/_game');
 // Play H5P games
 routes.add('play', `/${langParam}/games/play/:id`, 'games/_play');
 
-// Browse the books
-routes.add('browseGames', `/${langParam}/games/browse`, 'games/browse');
-
 // Read book
 routes.add(
   'read',
